Add new images and pay respects to T.O.P. etc
Add new paragraphs of body text to explain the reason for the website
and to give credit to the recipe's authors.
Add information and links to give credit to "The Odin Project" for
getting me to this point, and too allrecipes.com for having recipes
for me to use in this section.